backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
sqlite3

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# AI Integration
zhipuai==4.0.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0

# Report generation
reportlab==4.1.0
fpdf==1.7.2

# CORS
python-multipart==0.0.6

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
